+++ config/routes.rb
@@ -16,6 +16,7 @@ class Sowing::Application
   use Sowing::Controllers::TemplatesController
   use Sowing::Controllers::GuidesController
   use Sowing::Controllers::SettingsController
+  use Sowing::Controllers::SynthController
   use Sowing::Controllers::PreviewController
   use Sowing::Controllers::ApiController